Add files via upload
Shows a little bit of static after 1 iteration. Noticeable amount of noise after 10. Completely garbled after 100. Due to such a wild amount of difference, might want to add an option to the randomizer button that asks how "random" they want the colors (that is, choose a number that loops, maybe up to a cap). Probably a similar sort of feature will be wanted for the polarizer.