Add files via upload
Added a thumbnail function to Abstractinator. 

Added an "abstractinate" function as well. See documentation for how it works. Also check the constructor for an instance of its use, as well as how thumbnail is used (adjust width and height to your liking).

For that reason, also updated the Abstractinator to update the list when it's constructed and when the image is abstracted. Need to add ability to add a new imported image to the list.

Changed AbstractImage to include the thumbnail version of itself as well.